# internal/ports/http/todo/openapi.yml
openapi: '3.0.3'
info:
  title: Todo API
  version: '1.0.0'
  description: API for managing todo items
servers:
  - url: http://localhost:8080/api

paths:
  /todos:
    get:
      operationId: getTodos
      description: 'Get all todos'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
    post:
      operationId: createTodo
      description: 'Create a new todo'
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /todo/{id}:
    get:
      operationId: getTodo
      description: 'Get a todo by id'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
    put:
      operationId: updateTodo
      description: 'Update a todo'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
    delete:
      operationId: deleteTodo
      description: 'Delete a todo'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error

components:
  schemas:
    Todo:
      type: object
      required:
        - id
        - title
        - description
        - completed
        - createdAt
        - lastUpdatedAt
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastUpdatedAt:
          type: string
          format: date-time
    CreateTodoDto:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
        description:
          type: string
    UpdateTodoDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean